backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic-settings
